Implemented proper single-writer locking
There is a shared NSRecursiveLock used by all CBLForestDB instances on
the same file (path). The lock is acquired while in a transaction. This
enforces that only a single thread can write to the database.
Note that it isn't necessary to lock out readers during a transaction,
since the database is append-only.