Improved previous index optimization
Getting the old stored row to compare with is expensive; so now I compute
a 32-bit hash of the values and store that along with the keys. Then if
the old and new hash are different I know for _sure_ that the old row
is different, so there's no need to fetch and compare it.