Implemented database encryption using AES-256
Contains a kludgy workaround for compaction: it detects the filename
suffix added by the compactor for the temporary file, and encrypts that
file with the same key as the base file.
A better solution to this will require some changes to ForestDB, but we
haven't yet worked out what those would be.