Implemented fdb_copy_open_file as workaround for compaction issues
Couchbase Lite needs to create a copy of a database file with a different
encryption key, as part of its re-keying process. My earlier attempt to
do that using fdb_compact didn't work, due to the semantics of
compaction. So instead I've written a C function that simply writes a
copy of an already-open file with a new encryption key.
Database::copyToFile now calls the new function.