Fixed a nasty string encoding bug in Collatable
The kCharPriority table didn't include values for control characters
(and 0x7F), so they mapped to 0. But 0 is reserved to mean end-of-string
so when reading the string it would be truncated at that spot, causing
the rest of parsing to fail...

Bumped the version number of map-reduce indexes, to force all existing
indexes to be deleted and rebuilt correctly.

Also fixed a minor glitch in CollatableReader::dumpTo when dumping an
empty (zero-length) value.

For couchbase/couchbase-lite-ios/#899