Fixes & tweaks for the Windows-compatibility changes
The change to LogInternal.hh is critical -- the previous change
somehow broke the argument passing in log calls, resulting in crashes
inside vasprintf. I think the syntax for vararg macros must be slightly
different in MSVC.

The other notable change is that the encryption API is exposed in
Database.hh whether or not CBFOREST_ENCRYPTION is defined. I didn't
want the client to have to worry about having the right setting of
that macro. Instead, attempting to open a database with encryption will
throw an exception if the library isn't built with encryption support.