Eliminate per-platform targets, support AppleTV
[Cherry-picked from 48d8bdc]

By changing the SUPPORTED_PLATFORMS build setting to a list of all
platforms, I made each target buildable for Mac OS, iOS and tvOS, so
we only need one CBForest target, one forestdb target, etc.

Use the scheme destination pop-up, or the -destination arg of
xcodebuild, to specify which platform to build for.