C API: Got c4doc_insertRevisionWithHistory to work (API CHANGE!)
Sorry, it was broken before, and I didn't know because I forgot to
write a unit test for it. It works now and has a test.

I removed the redundant revID parameter -- the revID of the new revision
should be given as the first item in the history[] array. I've updated
the Java binding accordingly.