Java: Support for simultaneous indexing of multiple views
Broke the indexing API out into a new class, Indexer.
An Indexer is created with an array of views to be indexed.

Before calling a view's map function, call Indexer.shouldIndex to check
whether that view needs that sequence to be indexed; if not, skip it.

When emitting key/value pairs, you'll need to specify which view
they're for.

Fixes #47