JNI: Fixed local-refs overflow in Document.insertRevisionWithHistory
This function can use an arbitrary number of local JNI refs (one for
each item in the history array), so request capacity to handle them.
Unfortunately the maximum local-ref capacity is limited, and the actual
limit is unknown (implementation-specific but not available via API),
so we'll restrict ourselves to 200 (MaxLocalRefsToUse.) Android's limit
is 512.

If we have too many strings, copy the excess ones to the heap and
immediately release their JNI refs during the iteration.

Fixes couchbase/couchbase-lite-android#782