Removed assertions from RefCounted
It's dangerous for RefCounted to use CBFAssert because it throws
exceptions. Log warnings instead. In the destructor I chose to call
abort() because continuing after deleting an object that still has
references could cause really bad stuff like heap corruption. The
abort() call will be easier to recognize in a crash log than the rather
weird effects of throwing from the destructor.