Refactoring the C++ classes behind the C API (c4Database etc.)
Making these classes more robust and able to be used without the C API
functions, so they can serve as the top layer of the C++ API.

Also added `tryCatch` functions to wrap the exception handling: they
take a callback and invoke it which catching any exceptions and reporting
them in the outError parameter.

Also made all the C API functions `noexcept` so that (a) they can't
possibly throw exceptions; the process will abort instead, which is
easier to debug; and (b) the compiler won't emit the exception tables
for those functions, which bloat the code size.