Cannot define C4Slice differently externally vs internally
We used to declare C4Slice as a typedef for slice when compiling the
C API function implementations, as a convenience to let us use the full
slice API. Unfortunately this runs into trouble with some calling
conventions that pass plain-old structs differently from objects that
have constructors — the caller passes the C4Slice as a struct, but the
implementation was expecting it as an object, so it gets garbage.
(Apparently this happens with x86/32.)

Now c4Internal.hh defines C4Slice as a plain struct with no ctors.