c4db_delete shouldn’t fail if the refcount is >1
It won’t always be possible for platform bindings to free all C4 refs in
the right order, so the C API needs to be forgiving. It’s already ok to
close a db when it has a refcount > 1, so it’s actually OK to delete it
too. Removed the obsolete failure condition.

Also corrected the doc-comment to say that the handle is _not_ freed.