C API: Added c4blob_getFilePath()
This is to allow platform bindings to read the file efficiently by
memory-mapping it, or by returning a platform-specific file input stream
object. However, they need to have a fallback if this fails, because it
will fail if the database is encrypted, or if we change the implemen-
tation of blobs in the future.