Optimization for RevTree::insert
Most of the time the new rev being inserted becomes the default rev, so
if we insert it at the beginning of the _revs array, the array remains
sorted. I noticed the sorting was taking up significant time while
saving docs with lots of revs. This optimization sped up that test by
about 6%.