C4Error can now store custom error messages
An error can now have a custom message, not just a generic one based on
its domain/code. This is very useful for query parser errors because
there are a lot of possible messages and they may refer to specific
text in the JSON.

I had to add another field to C4Error; this may affect bindings!
It’s a uint32_t called internal_info which should be left alone.
c4error_getMessage() knows how to use it.