Don’t assign outgoing Messages numbers till they’re being queued
If the message number is assigned when the object is created, then
messages can be queued out of numeric order, and their first frames
will be sent out of order too, which is a protocol violation.

Also extended BLIPTest to create 100 Echoer actors and to detect when
they’re (almost) done. It doesn’t actually wait for their final responses
to be sent.