Fixed a nasty slice-related crasher in _onWebSocketMessage
Moving the `buf` pointer of an alloc_slice causes Bad Stuff to happen
when the destructor runs. Fixed _onWebSocketMessage to not do this.
I also updated the `slice` API in Fleece to prevent this from happening,
which required some updates in other places where I was (safely)
messing with the fields of an alloc_slice.