Database upgrading: don’t modify attachment metadata
I had second thoughts about modifying attachment metadata to add the
new “_cbltype”:”blob” property. This revision disables that (but makes
it possible to re-enable it with a #define.)

This does mean that attachments in imported databases will _not_ be
recognized as blobs by CBL 2.0 ... I think that will need to be fixed
at the public-API layer.