Refactored WebSocketImpl so the framing is optional
Some platforms already have a WebSocket implementation, but it’s still
valuable to layer it on top of WebSocketImpl. So the constructor now
takes a boolean ‘framing’ parameter; setting this to false means that
the instance will just pass the messages straight through to the
abstract Provider calls, and it won’t handle the close protocol.