Allow C4SocketFactory to provide a full WebSocket implementation
If the registered C4SocketFactory’s providesWebSockets field is true,
the factory provides a WebSocket implementation (not just a TCP socket)
and C4Socket will send it the WebSocket messages directly without
framing them, and will not un-frame received messages. It’s also assumed
that the factory takes care of the closing handshake.