Better fix for ThreadedMailbox::enqueueAfter
The mailbox shouldn’t keep its Actor alive until the delayed event fires;
that causes confusion because the actor may think it was finished. (And
it’s different than the behavior of GCD.)
Instead, use a tiny proxy object so that the delayed event can be safely
ignored if the actor has already been deleted.