Implemented lightweight logging
LogEncoder, LogDecoder implement efficient logging to a compact binary format.
The format is documented at the bottom of LogEncoder.cc.

LogDomain::writeEncodedLogsTo() will echo all logs to a file, in this format.
You should probably also set LogDomain::Callback to null, so everything doesn't
get written to stderr too.
The public API for this is c4log_writeToBinaryFile().

There's a new tool called 'litecorelog' that takes a log file path as argument
and writes it in readable form to stdout.

Fixes #108