Fixed a crash when BLIP connection fails
If the line 'webSocket->connect(this)' throws an exception, the BLIP
object's destructor will be called while unwinding the stack. If the
object has already been retained, the destructor will fail an assertion
because it's being destructed with a nonzero refcount.
Fixed by moving the retain call to the end of the constructor.