DataFile's fleeceAccessor and documentKeys should be initialized earlier
These properties are used by the SQLite Fleece functions registered by
SQLiteDataFile, so they need to be set before those functions are
registered. This broke in the previous commit, where I moved the
registration to the end of the reopen() method. :(
What I've done now is put those properties into the Options struct so
they're set as soon as the DataFile is constructed. This is a cleaner
design anyway, IMHO.