Added c4doc_update(), a more efficient way to update/delete a document
If you already have a C4Document, calling c4doc_update() will be a lot
faster than calling c4doc_put(), because it will use the existing
document instead of reading it from the database.

If your copy of the document is out of date (its sequence number is no
longer current) the call will fail with error kC4ErrorConflict, and
you'll need to read a new copy of the document and deal with conflict
resolution.

Also added companion function c4doc_create() for the case of creating a
document. It's basically just a wrapper around c4doc_put().