Cleaned up RevTree to avoid moving Rev objects around
It used to keep a vector<Rev>, so when inserting/deleting/sorting the
Revs, pointers to them would be invalidated. Plus, moving the objects
around is expensive.
Now it keeps the Revs themselves in a deque, simply because deque never
invalidates pointers to items even when it grows. The main structure is
now a vector<Rev*>, which is much faster to reorder and doesn't
invalidate anything.