Fixed: Puller messes up docs by encoding them without SharedKeys
The Fleece encoding created by IncomingRev can't be added to the db
because it doesn't use SharedKeys. We still need to create it, though,
so that IncomingRev can look for doc metadata and blobs. So I just
made DBWorker re-encode it using SharedKeys. We'll lose a bit of
performance but hopefully not too much.

Fixes #156