Create SQL index on sequence before running a FTS query
The SQL queries generated for the FTS match operator involve a join
against the `sequence` column, so they need an index on that column in
order to be fast. That index is created lazily, and unfortunately was
only created by a by-sequence all-docs enumeration.
This commit causes the index to be created if necessary when any FTS
query is compiled. It speeds up FTS queries a *LOT*.