Ensure conflicting revisions never become the current revision
Added revision flag kIsConflict.
This flag is automatically set on a revision that creates a branch, i.e.
one whose parent revision is not a leaf, and all its descendents.
A revision with this flag can never be the current revision.
However, if there are no other non-deleted leaf revisions, the flag is
removed. (This allows a conflict to be resolved by deleting the other
branch, making the conflicting branch the main one.)

Fixes #164