Fixed checkpoint recovery when local checkpoint is missing
If there's no local checkpoint, we still have to get the remote one,
because if one exists, we need to get its revID to be able to save a
new checkpoint. (Without this fix, saving the checkpoint results in a
conflict error.)
But we don't have to wait for the remote checkpoint, we can start
replicating right away. (This speeds up first-time replications.)