Replicator: More no-conflict work, updated 'proposeChanges' schema
'proposeChanges' now sends the current revID as well as the last known
server revID. That way if the server already has this revision, it
can say so by returning a 304 code and the pusher will skip it.

Without this change, recovering from lost checkpoints is very
expensive because the replicator will be sending mostly revisions the
server already has, but it won't necessarily know the server revID
(because that's often computed from the checkpoint's local sequence),
leading to tons of conflicts.