Query: Detect illegal usage of MATCH and return a clear error
Turns out SQLite only allows an FTS MATCH expression as the value of a
WHERE clause, or in an AND that's the value of a WHERE. In other cases
the query compiles but results in a SQLite error 1 when run.

Oh, and also you can only have one MATCH expression per FTS index.

Until we can figure out what to do about this, I've added code to detect
illegal usage and return a clear error from the query parser.

Fixes #166 (not really, but the error is more clear now)