Fixed RevTree bug that could cause hang or crash after c4doc_update call
When copying a RevTree object, the copied Rev objects' parent pointers
were left pointing into the original RevTree...
The only thing that copies a RevTreeis c4doc_update(), and the one
unit test for it didn't create a doc with enough revisions, so there
were no parent pointers. As soon as I extended the unit test to create
more revs, the address sanitizer caught it.

Fixes #185