Fixed: Push replicator wasn't sending histories of revisions to SG
SG's response to a "changes" message doesn't include a "maxHistory"
property, and LiteCore was interpreting that as a value of 0, which
meant no history at all got sent with revisions. That meant the pushed
revisions got added as detached leaves if the document already existed,
i.e. conflicts, and in particular a deletion would never show up on
SG because it would never be the current revision.

Fixed by changing the pusher to interpret a missing "maxHistory"
property as defaulting to 20 (kDefaultMaxHistory, in Pusher.hh.)
Also changed the minimum allowable value to 1 instead of 0, since 0 is
pretty useless.

Fixes #197