Fixed puller not stopping/idling, if it gets an error on a revision
Was waiting for _requestedSequences to become empty; but that will never
happen if one of those revs fails to download, so it's not useful for
this purpose. (Those sequences need to stay in the set because it's
used to update the checkpoint.)

Renamed it to _missingSequences, for clarity.
Added a counter _pendingRevMessages to do the tracking.

Fixes #179