Statically link SQLite
LiteCore now compiles the copy of SQLite in vendor/SQLiteCpp/sqlite3/
and statically links it, instead of using the system dylib.

I had to make a small change to Error.cc because it was calling
sqlite3_errstr(), but this isn't available anymore outside of the
LiteCore dylib (and Error.cc gets used in the separate REST dylib.)
So instead, it now calls c4error_getMessage in this situation.

To implement the above, there is now a LITECORE_IMPL preprocessor flag
that's only defined when compiling the LiteCore library itself.

Fixes #208