litecp can now read/write directories of .json files
A source or destination that ends with a "/" is interpreted as a
directory path.
If this is a source, litecp will read all files with
a ".json" suffix as documents, using the filename minus the suffix as
the default docID.
If this is a destination, litecp will create a file for every document,
whose filename is the docID with ".json" appended. Any docID that begins
with "." or contains the path separator character causes an error and
will not be copied.