Fixed assertion failure reading encrypted blobs from a stream
If the read size is larger than the blob size, EncryptedStream would
fail an assertion. The unit tests didn't discover this because
(a) they only tested small reads;
(b) the read test called c4stream_getLength() before reading data,
    which coincidentally avoided the situation that triggered the bug.

Fixed the bug and updated the unit test to trigger the situation that
caused it.