Rekey blobs when rekeying a database
c4db_rekey wasn't doing anything to the BlobStore, so the blobs were
left with the old encryption (or lack of it.) This of course made them
unreadable after the database was reopened.

The BlobStore is now rekeyed as part of the database rekeying:
* Create new BlobStore in a new directory, with the new encryption
* Export all the blobs to the new store
* Rekey the database itself
* Move aside the old BlobStore
* Move the new BlobStore into place
* Update the BlobStore's encryption key (since it now has the new files)
* Delete the old BlobStore
If any step fails, it rolls back to the original state of affairs.

Fixes #234