cblite tool: Added "revs" subcommand
The "revs" subcommand displays the revision tree of a document,
including every revision's flags and whether it has a body.

Also, a "--" flag can now be used to terminate the list of flags; this
is necessary if the first non-flag argument happens to start with a "-"
character! For example, "revs -my-doc-id-" won't work, but
"revs -- -my-doc-id-" will.