Don't warn about unbound query param if it starts with "opt_"
Sometimes it's OK for a query parameter to be unbound; CBLPredicateQuery
has such a case with its 'limit' and 'offset' parameters. These were
generating a lot of warnings from LiteCore.
I made it so params whose names start with "opt_" will not trigger this
warning.

Also, the JSON query will be logged to `SQL` as well as the resulting
SQL, for easier troubleshooting.