Improved sanity checking of RefCounted class in debug builds
It now detects race conditions where one thread (correctly) releases
the last reference and is destructing the object, while simultaneously
another thread (incorrectly) retains the object.
This required a hack, setting the initial refCount to a weird illegal
value instead of 0, because otherwise the initial retain of a brand new
object looks the same as this bug, since the refCount goes from 0 to 1.
But this hack only exists in debug builds.