Fixed replicator object leak when continuous replication is stopped
Race condition in DBWorker where the call to _getChanges happens *after*
_closeConnection. The work _getChanges does isn't just pointless; it
also sets _pusher, which creates a reference cycle that never gets
cleaned up (because the cleanup of it is in _closeConnection, which
already happened, D'ohhh!)

The fix is to make _getChanges return without doing anything if the
connection is closed. The rest of the code here is basically just extra
replicator logging, _plus_ a useful C function c4_dumpInstances()
which logs the exact objects that are extant in memory, with their
addresses and class names. (Only works in debug builds, though.)