Fixes a build fail seen on Ubuntu 17.10 -- some libc++ header looks for <xlocale.h> and doesn't find it.
This appears an issue seen also by some other projects in recent history based on searching around, and from the looks of it is dependent on libc++ version. Resolved by adding another semi-fixed include path (relative to LIBCXX_INCLUDE) with an associated comment. I'm not sure if there is a cleaner solution?