Fixed invalid-iterator bug in SequenceTracker
Windows was crashing with an assertion failure due to passing an invalid
iterator to a std::list. Fixed.

There were no visible effects from this bug on Mac, probably due to lack
of assertions in libc++. (I tried enabling full debug mode in libc++
but was unable to resolve some link errors.)

Fixes #292