LogEncoder now auto-flushes after 1 second
The binary log encoder buffers its output in memory, and writes it to
disk when the buffer reaches 64kbytes. It also used to write after a
log call if it had been 5 seconds since the last write. However this
meant that the last few seconds of logs were usually not written.

I added a timer so the logs will be written after 1 second of
inactivity. This means there should be a max 1 second latency before a
log message is written to disk.