Replicator: Partial fix for multiple-remotes issues
DBWorker shouldn't skip revisions that are 'foreign' until the Pusher
knows for sure that the peer requires noConflicts mode (proposeChanges).

This fixes certain problems with multiple remotes, i.e. pulling from
one remote and then pushing to a different one. But it's not a full fix;
that's still to come...

For #320