Assign UUIDs when db is first created
Until now, UUIDs were generated on demand. Unfortunately that meant that
getting the UUIDs of a read-only database might fail, if they hadn't
been accessed before. Now we generate them when the database is first
created.

Also, we were mistakenly creating 256-bit UUIDs when the correct size is
128-bit. Fixed that. This will invalidate saved replication checkpoints
for existing databases.