Made WebSocket ref-counted to improve memory management
Unfortunately this means it's no longer possible to inherit from both
WebSocket and Actor, since both inherit from RefCounted: that would
result in a weird chimera with two ref-counts, either of which would
delete the object when it reached zero.

So in MockWebSocket and LoopbackWebSocket I had to split the Actor code
out into a separate class (Driver), and the WebSocket has a Driver
instance associated with it.

For couchbase/couchbase-lite-core#353