Added replicator heartbeat (keep-alive)
The replicator sends WebSocket PING messages every 5 minutes, to prevent
overeager middleware from closing an idle connection, and to detect
a dangling socket.

The interval can be customized by setting the replicator option
kC4ReplicatorHeartbeatInterval to a number (units are seconds.) A value
of zero disables the heartbeat entirely.

Fixes #356