Fix errors saving checkpoints from SG-Accel
SG-Accel sequence IDs are strings not numbers. The Puller was not
quoting them; this caused errors when saving the checkpoint because its
payload contained an unquoted string making it invalid JSON.

In turn, an error saving the checkpoint forced the Replicator to keep
its status as Busy forever.

Fixes #373 (I think; can't reproduce as I don't have SG-Accel set up)