Use different checkpoint IDs for different channel/filter/docIDs settings
The checkpoint ID is now derived from the channels, filter, and docIDs
options, as well as the remote URL and local DB UUID.
Added a test for this.
Fixed a longstanding C4Test issue: the kBody string is actually invalid
JSON, because numbers with excess leading zeros are not allowed. Noticed
this because it blew up the JSON-to-Fleece conversion in this test.

Fixes #386