Suppress harmless SQLite "file unlinked while open" warning
It happens when SQLite defers closing a sqlite3 handle because a
statement is still open, and then we delete the database file, and then
the statement closes, which lets SQLite finally close the file, and it
notices the file's been deleted.

This scenario is hard to avoid in a Java client that uses finalizers
to free queries. And we don't care what happens to the database file
because it's been deleted already. So turning off this warning prevents
confusion and worry.

I added a unit test of this situation. To make sure the warning is gone,
I made LiteCoreTest track the number of warnings logged. We can make
use of this in the future to have all tests fail if they log any warnings.

For #381