Small tweak to yesterday's zombie-db fix
We don't need to set the NO_CKPT_ON_CLOSE flag unless the database still
has open statements and will become a zombie. So we can move that code
into the `if (!_sqlDb->closeUnlessStatementsOpen())` block that handles
that case.

(This isn't going to affect what the code does, since the
`if (_sqlDb->getHandle() != NULL)` test effectively does the same thing.
But it's clearer this way. I've also added some lengthy comments
explaining what's going on.)

For #381