Avoid pushing conflicts to the server
Use `proposeChanges` even if the server isn't in no-conflicts mode; this
helps prevent pushing a change that creates a conflict. To make 100%
sure we don't create a conflict, set a new `noconflicts` property in
the `rev` request.
These changes can be disabled by setting a new replicator option
`outgoingConflicts` to true.
Renamed the existing option `noConflicts` to `noIncomingConflicts` to
avoid confusion.

NOTE: This requires a small change in Sync Gateway to support the above
changes on its end.

Fixes #181