Xcode: Added no-op "Debug-EE" and "Release-EE" configs
For compatibility with couchbase-lite-core's Xcode project. Turns out
if you build with custom configurations, subprojects need to have ones
with the same names or else you can get the wrong config.
To be clear: this has no effect on this project. Debug-EE is identical
to Debug, and same for Release (and they should remain so.)