Fix for saving checkpoint on 1st-time replication
When there's no local checkpoint, we start replicating without waiting
for the remote checkpoint to arrive. That's good. But if we try to save
the checkpoint before the remote checkpoint arrives, we don't know the
revID to use, so if there's a remote checkpoint we'll get a conflict.

This commit makes saveCheckpoint wait until the remote checkpoing has
arrived.