Reject revIDs with uppercase hex digits
LiteCore stores revIDs in a binary form that takes up half the space.
But when regenerating a revID (for the API or replication) it emits the
hex digest as lowercase. That means that an incoming revID with
uppercase digits will not be preserved; when it's fetched from the DB
it will be lowercase, causing problems.

For #448