Semi-fix for bug pushing to a db when checkpoints don't match
If the remote checkpoint doesn't match the local one (or is missing),
we can't trust the current-remote-revision metadata in any documents.
Added a flag to make the DBWorker skip it.

This isn't an ideal fix, because if the push replication is stopped
before it catches up, then some metadata is still inaccurate. When
the replication starts up it will find matching checkpoints and won't
know to ignore the metadata (much less which of it to ignore.)

I can imagine better fixes, but they require larger changes than I think
we can make this close to a release.

Fixes #453