Distinguish between deletions and resolved conflicts
Added a new revision flag kClosed, which is added to the tombstone
revision that closes a losing conflicting branch. Revisions with this
flag are even lower priority than regular deletions, so will never be
the current revision.

c4doc_resolveConflict() adds kClosed to the tombstone that closes the
losing branch.

Extended the "Incoming Deletion Conflict" and "Local Deletion Conflict"
tests to update the document after resolving the conflict, so they
reflect the current state of affairs of the failing integration test.

For #433