Don't mark locally-created revs with kRevConflict
In c4doc_put, only add the kRevConflict flag to an inserted revision if
the remoteDB field is nonzero, i.e. if this revision comes from a remote
server.
This won't have any effect on clients, but allows tests to create
conflicting rev trees; see the upcoming "Server Conflict Branch-Switch".