Handle "branch switch" of server-side conflicted document
On a pull we get only the current revision of the server-side doc. If
SG allows conflicts, the doc on the server might be conflicted. Changes
might cause a revision on a different branch to become current.

LiteCore needs to detect and handle this, otherwise it looks like a new
conflict that needs to be resolved locally. When this happens a message
gets logged to 'Sync' starting "c4doc_put detected server-side branch-
switch..."

For couchbase/sync_gateway#3359