Descriptive error message trying to sync with SG 1.x
If the replicator gets a 404 response, it now checks the “Server:”
response header to see if the server is Sync Gateway 1.x; if so, it
changes the error to 501 Unimplemented with a descriptive message.

Also fixed a bug where Replicator::_onClose was replacing a previous
error status with a generic one.

Also made sure the `cblite cp` command’s error logging is visible.

Fixes #424