Updated BLIP: Another iOS 9 fix, for missing symbol deflatePending()
zlib's deflatePending() is only available starting in iOS 10 /
macOS 10.12, even though <zlib.h> claims it's available in iOS 8 /
macOS 10.10. This results in a crash calling that function on iOS 9.

Workaround is to test whether the symbol exists, and skip it if not.

I have now verified that unit tests pass on iOS 9 simulator.

Fixes couchbase/couchbase-lite-core#471