Fixed race conditions in LoopbackProvider
It was possible for it to receive a message before its connect()
method was called, in which case the message would be ignored. This
was sometimes causing the first BLIP message received to be lost,
which meant the next message would trigger a fatal out-of-sequence
error.

The flow of control between LoopbackWebSocket and MockWebSocket (and
their Drivers) was getting impossible to keep track of, so I refactored
the Loopback classes to be standalone, merging in the code inherited
from the Mock classes.

Then I added a state machine to keep track of the connection state.

Then I changed connect() to use a handshake, to ensure both sides have
notified their delegates they're connected before either side sends a
message. In other words, to prevent one side from receiving a message
from the peer before it's ready.

For couchbase/couchbase-lite-core#474