Stop c4doc_update() from clobbering remote-revID state
I had to replace the implementation of c4doc_update with one that just
uses c4doc_put, so it won’t be as fast. (It’ll do a read-modify-write
instead of just modify-write.) But it’s the only thing I can think of
that’s safe to do for 2.0.0.

Still needs a unit test.

Fixes #478