Fixed crashing bug in WebSocketProtocol
The 3rd party WebSocketProtocol class has a very odd behavior in that
its consume() method may modify bytes before the start of the input
buffer passed to it.  The library it’s from (uWebSockets) was aware
of this and reserved space before the start of the buffer; but we
were not. This behavior doesn’t occur often, but when it does it can
cause crashes or heap corruption.
I’ve added a workaround and tested it by artificially causing frames
to be split so as to trigger the behavior.
Fixes couchbase/couchbase-lite-core#531