Tweak the log pruning test
Apparently the delays are not long enough on *nix to guarantee that the “oldest” two files will be detected as such.  Now just delay 2 seconds after each of the first two logs to put the rest in a newer timestamp so that log10 and log11 are the ones to be deleted.