Added ‘parentMailbox’ facility for Actors
An Actor can optionally have a parentMailbox; only one Actor with the
same parentMailbox can be executing at once. This can help improve
performance by limiting concurrency.
So far this is only implemented for GCDMailbox, where it determines
the mailbox’s dispatch queue’s ‘target queue’. This helps reduce
the number of threads that get spawned by GCD.
It could be implemented for ThreadedMailbox, but I think it’s less
important because that implementation only creates a fixed-size
thread pool.
For #561