cblite tool: “cp” of local db files does a c4db_copy()
When the cblite tool is told to copy a local db to a local db path,
it now calls c4db_copy() instead of using a local-to-local
replication. This provides a fast way to copy a database file and
assign it a new UUID, which is necessary if it’s going to be used
with the replicator.
You can still force the local-replication behavior by using the
“—replicate” flag.