Fixed C4Database instance leak race condition
c4db_free() would behave as a no-op if the instance was in a
transaction.  This made sense back when it actually freed the
instance, but not when it's just decrementing the ref-count!

This bug sometimes got triggered in ReplicatorLoopbackTest if the
test code was updating the db after a first replication, and at the
same moment the DBWorker from the first replication was destructing.
Its c4db_free would do nothing, so the reference would be leaked.