Replicator eventing now includes flags, revID [API CHANGE]
Added two parameters to C4ReplicatorDocumentEndedCallback.

The flags include a new kRevPurged flag, which on pull indicates that
the doc was purged because it was removed from all channels.

Also overhauled how these notification callbacks are managed
internally: they now use the same Rev (now ReplicatedRev) objects as
during the workflow, and a Batcher to collect them.

Fixes #607