Pass C4BlobStore to C4PredictiveQuery callback [API CHANGE]
This allows the callback to handle a Couchbase Lite blob that’s
passed in the input as a dictionary (the document property) without
having gone through the BLOB query operator.

NOTE: I decided not to pass the C4Database, because that would make
it possible for a callback function to write to the database or run
another query, which would be disastrous since the callback is being
invoked from within a SQLite query already.

Fixes #654