Sanity-check the replication URL [NEW API]
c4repl_new now returns an error if any of these mistakes are made:
* The scheme is not ws: or was:
* The remote db name is missing or invalid (not all-lowercase)
* The hostname is empty
* The port is 0

There’s also a new public function c4repl_isValidRemote() that makes
these checks, in case Couchbase Lite wants to fail earlier, like in
the Replicator object’s constructor.

This is in response to several forum/support issues where someone
forgot to add the db name to the path, or something similar, and the
errors happened after the replicator started and were hard to diagnose.