Fix remoteAncestorIDs not getting updated by replicator
The Replicator could sometimes stop before its DBWorker had a chance
to call _markRevsSyncedNow and update all docs' remoteAncestorID.

I fixed this in the Batcher class in BLIP-Cpp. Batcher now lets its
parent Actor know that there is work to be done, by enqueuing a
message immediately when the 1st item is added.

I also fixed an issue I saw in Replicator, where it could decide it's
stopped even if it still has events to be processed.

Fixes #658 (I hope)