Fix dynamic destructor issue and correct invalid behavior
Two fixes are in this commit.  One change `sRegistry` to be a pointer so that it won't be automatically freed upon process exit (the reasoning for this is in the comments).  The other is that `unordered_map` does not automatically replace items and so the insert behavior was incorrect (it would ignore any subsequent insert with the same key).