Fixed push conflict errors w/rapidly-changing doc
If a doc is updated when a change to it was just pushed (in a
continuous replication), there’s a race condition where the
‘proposeChanges’ message for the new rev will give an older rev as
the server-side base rev, instead of the one that was just pushed.
This then causes a 409 error from the server, and the new rev won’t
be pushed.
Fixed this by ensuring that a new rev will not begin to be pushed
until after an earlier rev is completely finished. To do this I had
to move some logic from Pusher to DBWorker.

Fixes #669 (I think ... we don’t have a reproducible case in-house)