Fixed SequenceTracker crash “Assertion failed: i != observers.end()”
The SequenceTracker could garbage-collect an entry corresponding to a
document that still had observers. If so, removing those observers
would crash because the entry wouldn’t be found.

(Customer issue, that I reproduced by altering a unit test.)

Fixed by having it move the entry to the idle list instead of deleting
it.

Fixes #693