More thorough doc body validation in debug builds
Database::validateRevisionBody() now checks for dict keys that
should have been encoded as shared-keys but weren’t. If such keys
got added to the database they’d likely cause problems, such as
later not appearing in queries.

Fixed some unit tests that were violating this check.

Unfortunately I had to change the value of kFleeceBody: to be a
constant it can’t involve a per-database shared key, so I had to
make the key it uses non-shareable by adding a `*`. I say “unfor-
tunately” because that changes the revIDs of any revisions using
that body. So I had to update tests that checked those.