Optimized pulling deltas
Deltas are now applied/expanded during the call to c4doc_put().
This saves an extra trip to the db to read the source revision.

This can’t be done in all cases: if the delta indicates that the
revision adds or alters a blob, or if the replicator has a
validation callback, the delta has to be applied earlier as it used
to be.