Upgrader: Skip “_design/…” docIDs
In 1.x (and CouchDB), this form of docID was allowed even though it
starts with a `_`. But 2.x doesn’t allow _any_ doc IDs that start
with `_`, so the Upgrader fails.
I added a check to detect and skip such docs. These weren’t “real”
docs anyway, they were used to store JavaScript map and reduce
functions, which are useless in 2.x anyway.

Fixes #714