Fixed delta sync when pull filter present
Applying a delta was failing when there was a pullValidator present,
and the delta JSON contained a new doc property not yet in the DB’s
SharedKeys.

In this situation the DBWorker now re-encodes the current rev body so
it doesn’t use the DB’s SharedKeys, before applying the delta to it.

None of the unit tests triggered this, because the delta tests didn’t
use a pullValidator (or vice versa.) I’ve remedied that.

Fixes #732